refactor(spam): move softbanMember to effects/discordSdk
The helper is not spam-specific — it's a Discord moderation primitive
(ban + unban to wipe recent messages). Moving it next to the other
Effect-wrapped SDK helpers makes it reusable by any future caller
outside the spam pipeline (manual softban command, raid response, etc.).

Matches the file's conventions: typed DiscordApiError on failure,
discord.softbanMember span name, attributes on the span. The unban-
failure case still propagates as DiscordApiError(operation=
"softbanMember.unban") so the caller can distinguish it from a failed
ban, AND is logged at error level inside the helper via tapError so
the "user left banned" operational incident is never lost even if a
caller forgets to handle it. Service tag changes from "SpamResponse"
to "Discord" to reflect the helper's new home.

Callers in spamResponseHandler.ts now wrap softbanMember with catchTag
("DiscordApiError", ...) and log at the spam-pipeline-appropriate level.
